==> project-files/03-import.ipynb <==
import requests
import json

# Configurações
auth_url = 'https://seu-endpoint-dadosfera.com/auth/signin'
upload_url = 'https://seu-endpoint-dadosfera.com/upload/json'
usuario = 'user@example.com'
password = 'yourPassword'
file_path = 'dataset/linhas_validas_com_features.json'

# Função para autenticar e obter o token
def get_auth_token(email, password):
    response = requests.post(auth_url, json={"usuario": usuario, "password": password})
    
    if response.status_code == 200:
        data = response.json()
        return data.get('token')
    else:
        print("Falha na autenticação:", response.status_code, response.text)
        return None

# Função para enviar o arquivo JSON
def upload_file(token, file_path):
    headers = {
        'Authorization': f'Bearer {token}'
    }
    
    with open(file_path, 'rb') as file:
        files = {'file': file}
        response = requests.post(upload_url, headers=headers, files=files)
        
        if response.status_code == 200:
            data = response.json()
            print("Arquivo enviado com sucesso!")
            print("Detalhes do Arquivo:", data)
        else:
            print("Falha no upload do arquivo:", response.status_code, response.text)

# Execução do script
token = get_auth_token(email, password)
if token:
    upload_file(token, file_path)
